more clippy, move stuff into tokio_util
