+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod rpc_protocol;
 mod subnet;
 mod tls;
 mod util;
+mod tokio_util;
 
 pub use tls::{Keypair, PeerId, PeerIdError, PublicKey, SecretKey, Signature};
 pub use util::Hash;
